+++ package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/user-event": "^12.1.3",
     "axios": "^0.21.1",
     "dotenv": "^8.2.0",
@@ -26,6 +25,7 @@
     "lint:fix": "eslint ./src --ext .js,.jsx --fix"
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^10.4.9",
     "eslint-config-airbnb": "^18.2.0",
     "eslint-config-prettier": "^6.11.0",
