+++ src/apis/youtube.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-const KEY = process.env.REACT_APP_YOUTUBE_API_KEY;
+const KEY = process.env.REACT_APP_YOUTUBE_API_KEY_ALT;
 
 export default axios.create({
     baseURL: 'https://www.googleapis.com/youtube/v3/',
