--- src/components/FavSmallVideo/index.js
@@ -1,1 +1,0 @@
-export { default } from './FavSmallVideo.component';
