--- src/components/FavouriteCards/index.js
@@ -1,1 +1,0 @@
-export { default } from './FavouriteCards.component';
