--- src/components/FavouriteCards/styled.js
@@ -1,14 +1,0 @@
-import styled from 'styled-components';
-
-const CardsContainer = styled.div`
-    display: flex;
-    flex-wrap: wrap;
-    margin: 10px auto;
-    padding: 10px;
-`
-
-const Styled = {
-    CardsContainer
-}
-
-export default Styled
