--- src/components/FavouriteVideoView/index.js
@@ -1,1 +1,0 @@
-export { default } from './FavouriteVideoView.component';
