+++ src/components/SideVideos/styled.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+const Videos = styled.div `
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+    @media (min-width: 900px){
+        width: 30%;
+    }
+`;
+
+const Styled = {
+    Videos
+}
+
+export default Styled;
