--- src/components/VideoView/RelatedVideos/RelatedVideo/index.js
@@ -1,1 +1,0 @@
-export { default } from './RelatedVideo.component';
