--- src/components/VideoView/RelatedVideos/index.js
@@ -1,1 +1,0 @@
-export {default} from './RelatedVideos.component';
