--- src/components/VideoView/index.js
@@ -1,1 +1,0 @@
-export { default } from './VideoView.component';
