+++ src/pages/Favourite/Favourite.styles.css
@@ -1,9 +1,9 @@
-.videopage {
+.favouritepage {
   text-align: center;
   width: 100%;
 }
 
-.videopage h1 {
+.favouritepage h1 {
   font-size: 3rem;
   letter-spacing: -2px;
 }
