+++ src/pages/FavouriteVideo/FavouriteVideo.styles.css
@@ -1,9 +1,9 @@
-.videopage {
+.favouritevideopage {
   text-align: center;
   width: 100%;
 }
 
-.videopage h1 {
+.favouritevideopage h1 {
   font-size: 3rem;
   letter-spacing: -2px;
 }
