+++ src/pages/FavouriteVideo/index.js
@@ -1,0 +1,1 @@
+export { default } from './FavouriteVideo.page';
