+++ src/pages/FavouriteVideo/styled.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+const Container = styled.div `
+    width: 100%;
+    height: 100%;
+    padding: 1em;
+    gap: 1em;
+    @media (min-width: 900px){
+        display: flex;
+    }
+`;
+
+const Styled = {
+    Container
+}
+
+export default Styled;
