+++ src/pages/Video/Video.page.jsx
@@ -1,11 +1,30 @@
 import React from 'react';
-import VideoView from '../../components/VideoView';
 import './Video.styles.css';
+import Video from '../../components/Video';
+import SideVideos from '../../components/SideVideos';
+import Styled from './styled';
+import useRelatedVideos from '../../hooks/useRelatedVideos';
+import { useParams } from 'react-router-dom';
 
 const VideoPage = () => {
+    const params = useParams();
+    const videoId = params.videoId;
+    const [isLoading, videos] = useRelatedVideos(videoId);
+
+    const getVideos = () => {
+        const videoIds = videos?.map((video) => video.id.videoId);
+        return <SideVideos videoIds={videoIds} linkTo='/'/>
+    }
+
     return (
         <section className="videopage">
-            <VideoView />
+            <Styled.Container >
+                <Video />
+                {isLoading ? <p>Loading...</p> : 
+            (
+                getVideos()
+            )}
+            </Styled.Container>
         </section>
     );
